# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Contact Notes API",
      "description": "API for managing contacts and their notes",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "description": "User's unique username"
            },
            "password": {
              "type": "string",
              "description": "User's password"
            }
          },
          "required": ["username", "password"]
        },
        "Contact": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "Contact ID"
            },
            "name": {
              "type": "string",
              "description": "Contact name"
            },
            "email": {
              "type": "string",
              "description": "Contact email address"
            }
          }
        },
        "Note": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "Note ID"
            },
            "body": {
              "type": "string",
              "description": "Note content"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "description": "Creation timestamp"
            }
          }
        },
        "Error": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "description": "Error message"
            }
          }
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Register a new user",
          "security": [],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "User created successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "409": {
              "description": "Username already exists"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login and get access token",
          "security": [],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "access_token": {
                        "type": "string"
                      },
                      "refresh_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
      "/auth/refresh": {
        "post": {
          "summary": "Refresh access token",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Token refreshed",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "access_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Invalid token"
            }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "summary": "Logout and invalidate token",
          "responses": {
            "200": {
              "description": "Logout successful"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/contacts": {
        "get": {
          "summary": "Get all contacts",
          "responses": {
            "200": {
              "description": "A list of contacts",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "post": {
          "summary": "Create a new contact",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  },
                  "required": ["name"]
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Contact created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/contacts/{contact_id}": {
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the contact"
          }
        ],
        "get": {
          "summary": "Get a specific contact",
          "responses": {
            "200": {
              "description": "Contact details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            },
            "404": {
              "description": "Contact not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "put": {
          "summary": "Update a contact",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Contact updated",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            },
            "404": {
              "description": "Contact not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "delete": {
          "summary": "Delete a contact",
          "responses": {
            "200": {
              "description": "Contact deleted"
            },
            "404": {
              "description": "Contact not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/contacts/{contact_id}/notes": {
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the contact"
          }
        ],
        "get": {
          "summary": "Get all notes for a contact",
          "responses": {
            "200": {
              "description": "A list of notes",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Note"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Contact not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "post": {
          "summary": "Create a new note for a contact",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "body": {
                      "type": "string",
                      "description": "Note content"
                    },
                    "note_body": {
                      "type": "string",
                      "description": "Alternative field for note content"
                    },
                    "note_text": {
                      "type": "string",
                      "description": "Alternative field for note content"
                    }
                  },
                  "description": "At least one of the body fields must be provided"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Note created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Note"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "Contact not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
      "/contacts/{contact_id}/notes/{note_id}": {
        "parameters": [
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the contact"
          },
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the note"
          }
        ],
        "get": {
          "summary": "Get a specific note",
          "responses": {
            "200": {
              "description": "Note details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Note"
                  }
                }
              }
            },
            "404": {
              "description": "Note not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "put": {
          "summary": "Update a note",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "body": {
                      "type": "string",
                      "description": "Note content"
                    },
                    "note_body": {
                      "type": "string",
                      "description": "Alternative field for note content"
                    },
                    "note_text": {
                      "type": "string",
                      "description": "Alternative field for note content"
                    }
                  },
                  "description": "At least one of the body fields must be provided"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Note updated"
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "Note not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "delete": {
          "summary": "Delete a note",
          "responses": {
            "200": {
              "description": "Note deleted"
            },
            "404": {
              "description": "Note not found"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      }
    }
  }